Fix a segfault and assert handling
When having a FAT partition already resized to the maximum size and
trying to resize it again to maximum size, then fatresize segfaults
because of end_part being NULL. So check it for NULL and return from
snap_to_boundaries() in this case.

While debugging this, it has been noticed that FAT_ASSERT is
ineffective and the action is not executed with parted >= 2.4. So
add execution of the requested action and define DEBUG to make
PED_ASSERT effective. Also add an assertion for the start_part in
snap_to_boundaries().

Verified by testing and disassembly.
Fixes #6

Signed-off-by: Sebastian Parschauer <sparschauer@suse.de>