Allow setting maximum size by keyword "max"
Setting the maximum size requires two steps at the moment:
1. get the maximum size with the info option
2. set the maximum size from the info option

It would be much nicer to combine this. So introduce the keyword
"max" for the size. Use LLONG_MAX internally and reuse code parts
from the info option to set the actual maximum size. Update help
text and man page as well.

Looks like this:
$ fatresize -s max /dev/sdb1

Fixes #5

Signed-off-by: Sebastian Parschauer <sparschauer@suse.de>