get_device: call get_partnum with dev, not devname
By the time get_partnum() is called here, devname may have been modified
to it no longer ends with the partition number but instead is a
null-terminated device name.  That is, dev may still be '/dev/sda3' but
devname is now '/dev/sda'.

Additionally, the '-n' parameter hasn't been respected, with get_partnum()
consistently returning '1' regardless of the actual partition specified on
the command line.

Only attempt to deduce the partition number if it wasn't already
explicitly stated on the command line.  This would also be essential if
mmc devices should be supported in the future as they're frequently named
mmcblk0p1 where a partition name isn't an int but a string.

Signed-off-by: Joe MacDonald <joe_macdonald@mentor.com>